Computer_vision/requirements.txt
python-dotenv
Pillow
matplotlib
azure-core
azure-ai-vision
requests
